// types.ts
export enum Preset {
  Portraitify = 'Portraitify',
  DressRecolor = 'DressRecolor',
  BackgroundReplace = 'BackgroundReplace',
  ObjectRemoval = 'ObjectRemoval',
  ColorizeBW = 'ColorizeBW',
  ProductStage = 'ProductStage',
  RemoveBlemishes = 'RemoveBlemishes',
  VintageFilm = 'VintageFilm',
  HairHighlight = 'HairHighlight',
  TeethWhitening = 'TeethWhitening',
}